Refactor for compiling as a library for statically linking into other open source applications. Essentially, move main() into its own file.
